feat: add an example for BLOOM
